requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.1
django-address==0.1.5
django-anymail==0.9
enum34==1.1.6
geopy==1.11.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
six==1.10.0
whitenoise==3.3.0
